requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
discord==2.1.0
discord.py @ git+https://github.com/rapptz/discord.py@7cf3cd51a5d463ae5633ec2dd43d3e397b081876
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pycparser==2.21
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.13
vboxapi==1.0
yarl==1.8.2